DRAFT - Sample extension plugin
Provides a new Viewpoint with an optional but activated by default
layer. 
This layer contains a new tool whose visibility depends on the
synchonized state of the diagram. 

Activate the viewpoint, unsynchronize the diagram, the delete from
diagram is now enabled, as the new insert speaker tool. The tool propose
to create the views of the existing speakers who are not lareay on the
diagram (should improved to depends on the tracks)

I have also set the linkBubble mapping to synchronized, to have the
corresponding edges created automatically.